safari-534.53-branch/LayoutTests/platform/mac/accessibility/aria-alerts-expected.txt

alert role
alert dialog role
dialog role
This tests that each of these ARIA groups has the right role and subrole.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).role is 'AXRole: AXGroup'
PASS body.childAtIndex(0).subrole is 'AXSubrole: AXApplicationAlert'
PASS body.childAtIndex(1).role is 'AXRole: AXGroup'
PASS body.childAtIndex(1).subrole is 'AXSubrole: AXApplicationAlertDialog'
PASS body.childAtIndex(2).role is 'AXRole: AXGroup'
PASS body.childAtIndex(2).subrole is 'AXSubrole: AXApplicationDialog'
PASS successfullyParsed is true

TEST COMPLETE
